Move `RecursionPublicValues` to `sp1-hypercube` from `sp1-recursion-gnark-ffi`
