+++ crates/verifier/Cargo.toml
@@ -29,7 +29,6 @@ strum = { workspace = true, features = ["derive"] }
 sp1-hypercube = { workspace = true }
 sp1-primitives = { workspace = true }
 sp1-recursion-machine = { workspace = true, default-features = false }
-sp1-recursion-executor = { workspace = true, default-features = false }
 slop-algebra = { workspace = true }
 slop-challenger = { workspace = true }
 slop-primitives = { workspace = true }
